Move interchange to interfaces
